added misspelling and fixed a bug in numbers
numbers can now replace characters with random numbers instead of from the number map if given the -m option, bug in numbers was if a letter not in the map was chosen to be replaced it would be replaced with the last character replaced. i shudder to think about what would've happened if it chose to replace a character not in the map before any other character was replaced.